.emacs: try to fix errors in debian system el, browse-key-ring, bbdb .emacs_d/lisp/bhj-autoloads.el: autoloads update .emacs_d/lisp/bhj-defines.el: no need for bbdb de-duplicate bin/Linux/config/install-missing-pkgs: do not install missing pkgs if there are too many bin/Linux/config/upd_system: install texlive only on my own PC bin/Linux/emacs-install-packages: make sure ~/.emacs.d is there already bin/adb-fix-wifi-mac: print the wifi mac bin/adb-picture-to-weixin-chat: add some comments, add some waiting bin/after-co-ln-s.sh: fix ~/.cache created by update-host-ip. fix systemctl not found bin/do-remote-system-config: make sure the password is used; make sure root can connect bin/search-google: can search sf bin/search-google: search sf ticket bin/smbcp: remount the ~/smb if necessary

+++ bin/after-co-ln-s.sh
@@ -236,7 +236,7 @@ if test -x ~/src/github/private-config/after-co.sh; then
     ~/src/github/private-config/after-co.sh
 fi
 
-if test ! -e /etc/systemd/system/rc-local.service; then
+if which systemctl >/dev/null 2>&1 && test ! -e /etc/systemd/system/rc-local.service; then
     sudo cp ~/etc/systemd/system/rc-local.service /etc/systemd/system/rc-local.service
     sudo chmod a+X /etc/systemd/system/rc-local.service
     sudo systemctl --system daemon-reload
@@ -250,5 +250,7 @@ fi
 sudo mkdir -p /etc/acpi/local/
 
 sync-etc-files
+mkdir -p ~/.cache # just in case the following command will create
+                  # .cache with root permission.
 update-host-ip phone 192.168.15.244
 update-host-ip home 128.199.178.185
